Change SpdyFrameBuilder::length() to return size_t instead of int to avoid a bunch of static_casts in SpdyFramer.
Review URL: http://codereview.chromium.org/10162008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@133321 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/spdy/spdy_framer.cc
@@ -1052,7 +1052,7 @@ SpdySynStreamControlFrame* SpdyFramer::CreateSynStream(
   frame.WriteUInt8(priority << ((spdy_version_ < 3) ? 6 : 5));
   frame.WriteUInt8((spdy_version_ < 3) ? 0 : credential_slot);
   WriteHeaderBlock(&frame, headers);
-  DCHECK_EQ(static_cast<size_t>(frame.length()), frame_size);
+  DCHECK_EQ(frame.length(), frame_size);
 
   scoped_ptr<SpdySynStreamControlFrame> syn_frame(
       reinterpret_cast<SpdySynStreamControlFrame*>(frame.take()));
@@ -1085,7 +1085,7 @@ SpdySynReplyControlFrame* SpdyFramer::CreateSynReply(
     frame.WriteUInt16(0);  // Unused
   }
   WriteHeaderBlock(&frame, headers);
-  DCHECK_EQ(static_cast<size_t>(frame.length()), frame_size);
+  DCHECK_EQ(frame.length(), frame_size);
 
   scoped_ptr<SpdySynReplyControlFrame> reply_frame(
       reinterpret_cast<SpdySynReplyControlFrame*>(frame.take()));
@@ -1109,7 +1109,7 @@ SpdyRstStreamControlFrame* SpdyFramer::CreateRstStream(
                          frame_size);
   frame.WriteUInt32(stream_id);
   frame.WriteUInt32(status);
-  DCHECK_EQ(static_cast<size_t>(frame.length()), frame_size);
+  DCHECK_EQ(frame.length(), frame_size);
   return reinterpret_cast<SpdyRstStreamControlFrame*>(frame.take());
 }
 
@@ -1127,7 +1127,7 @@ SpdySettingsControlFrame* SpdyFramer::CreateSettings(
     frame.WriteUInt32(it->second.second);
     ++it;
   }
-  DCHECK_EQ(static_cast<size_t>(frame.length()), frame_size);
+  DCHECK_EQ(frame.length(), frame_size);
   return reinterpret_cast<SpdySettingsControlFrame*>(frame.take());
 }
 
@@ -1135,7 +1135,7 @@ SpdyPingControlFrame* SpdyFramer::CreatePingFrame(uint32 unique_id) const {
   size_t frame_size = SpdyPingControlFrame::size();
   SpdyFrameBuilder frame(PING, CONTROL_FLAG_NONE, spdy_version_, frame_size);
   frame.WriteUInt32(unique_id);
-  DCHECK_EQ(static_cast<size_t>(frame.length()), frame_size);
+  DCHECK_EQ(frame.length(), frame_size);
   return reinterpret_cast<SpdyPingControlFrame*>(frame.take());
 }
 
@@ -1154,7 +1154,7 @@ SpdyGoAwayControlFrame* SpdyFramer::CreateGoAway(
   if (protocol_version() >= 3) {
     frame.WriteUInt32(status);
   }
-  DCHECK_EQ(static_cast<size_t>(frame.length()), frame_size);
+  DCHECK_EQ(frame.length(), frame_size);
   return reinterpret_cast<SpdyGoAwayControlFrame*>(frame.take());
 }
 
@@ -1181,7 +1181,7 @@ SpdyHeadersControlFrame* SpdyFramer::CreateHeaders(
     frame.WriteUInt16(0);  // Unused
   }
   WriteHeaderBlock(&frame, headers);
-  DCHECK_EQ(static_cast<size_t>(frame.length()), frame_size);
+  DCHECK_EQ(frame.length(), frame_size);
 
   scoped_ptr<SpdyHeadersControlFrame> headers_frame(
       reinterpret_cast<SpdyHeadersControlFrame*>(frame.take()));
@@ -1206,7 +1206,7 @@ SpdyWindowUpdateControlFrame* SpdyFramer::CreateWindowUpdate(
                          frame_size);
   frame.WriteUInt32(stream_id);
   frame.WriteUInt32(delta_window_size);
-  DCHECK_EQ(static_cast<size_t>(frame.length()), frame_size);
+  DCHECK_EQ(frame.length(), frame_size);
   return reinterpret_cast<SpdyWindowUpdateControlFrame*>(frame.take());
 }
 
@@ -1235,7 +1235,7 @@ SpdyCredentialControlFrame* SpdyFramer::CreateCredentialFrame(
     frame.WriteUInt32(cert->length());
     frame.WriteBytes(cert->c_str(), cert->length());
   }
-  DCHECK_EQ(static_cast<size_t>(frame.length()), frame_size);
+  DCHECK_EQ(frame.length(), frame_size);
   return reinterpret_cast<SpdyCredentialControlFrame*>(frame.take());
 }
 
@@ -1247,7 +1247,7 @@ SpdyDataFrame* SpdyFramer::CreateDataFrame(SpdyStreamId stream_id,
   size_t frame_size = SpdyDataFrame::size() + len;
   SpdyFrameBuilder frame(stream_id, flags, frame_size);
   frame.WriteBytes(data, len);
-  DCHECK_EQ(static_cast<size_t>(frame.length()), frame_size);
+  DCHECK_EQ(frame.length(), frame_size);
   return reinterpret_cast<SpdyDataFrame*>(frame.take());
 }
